"Removed logo images, added Navbar and About components to App.js"
This commit removes the logo192.png and logo512.png files from the public directory. It also updates the App.js file by removing the logo import and header section, and adding a Navbar component and an About component. The new components are imported at the top of the file. The return statement in the App function now includes the Navbar and About components, wrapped in a fragment shorthand (<></>). The TextForm component has been commented out.